Create parent preferences controller class
