+++ config/locales/en.yml
@@ -1949,6 +1949,11 @@ en:
       flash changed: "Your password has been changed."
       flash token bad: "Did not find that token, check the URL maybe?"
   preferences:
+    preferences:
+      update:
+        failure: Couldn't update preferences.
+      update_success_flash:
+        message: Preferences updated.
     basic_preferences:
       show:
         title: My Preferences
@@ -1963,10 +1968,6 @@ en:
           light: Light
           dark: Dark
         save: Update Preferences
-      update:
-        failure: Couldn't update preferences.
-      update_success_flash:
-        message: Preferences updated.
   profiles:
     show:
       title: Edit Profile
